Meme Generator
This is a Meme Generator app build using react. I used an API that contains popular memes so I could use them with my app. Within this app, a user can type a message for the top and bottom of the meme image, then generate a new meme with that text.